resolve dependencies errors

+++ trainers.py
@@ -25,8 +25,9 @@ def mcc_scores(probs, Y_test, N=101):
     """
     thresholds = np.linspace(0, 1, N)
     mccscores = np.zeros(N)
+    Y_test = Y_test.astype(int)
     for i in range(N):
-        yp = (probs >= thresholds[i]).astype(np.int)
+        yp = (probs >= thresholds[i]).astype(int)
         mccscores[i] = matthews_corrcoef(Y_test, yp)
     return thresholds, mccscores
